requirements.txt
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
chromedriver==2.24.1
chromedriver-binary==86.0.4240.22.0
click==7.1.2
constantly==15.1.0
cryptography==3.2.1
cycler==0.10.0
dateparser==1.0.0
ffmpeg==1.4
flake8==3.8.3
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kiwisolver==1.2.0
matplotlib==3.3.1
mccabe==0.6.1
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
playsound==1.2.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybitflyer==0.1.9
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.13
requests==2.24.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
spark-parser==1.8.9
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
uncompyle==2.0.0
uncompyle6==3.7.4
urllib3==1.25.10
websocket-client==0.47.0
xdis==5.0.4
zope.interface==5.2.0
